pyproject.toml
[project]
name = "TROLoad"
version = "1.2.0"
description = "Pod to load data from Quicken into the TRO database."
authors = [{ name = "Jeff Smith", email = "jasmit@swbell.net" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.12,<3.13"
dependencies = [
    "config>=0.5.1",
    "openpyxl>=3.1.5",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.1",
]

[project.urls]
Homepage = "https://jasmit.github.io/TROLoad/"
Repository = "https://github.com/jasmit/TROLoad"
Documentation = "https://jasmit.github.io/TROLoad/"

[copyfiles]
python = [
    "/Users/jeff/devl/TRO/src/tro/python/categories_table.py"
]

[tool.uv]
dev-dependencies = [
    "autoflake>=2.3.1",
    "deptry>=0.20.0",
    "mkdocs>=1.6.1",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.19.0",
    "mypy>=0.991",
    "pre-commit>=2.20.0",
    "pytest>=7.2.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.0.235",
    "tox-uv>=1.11.3",
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.mypy]
files = ["src"]
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"
#
#  jasmit  Added following
ignore_missing_imports = "True"
#
#  jasmit  Switched following
disallow_untyped_defs = "False"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.format]
preview = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # ImportNotAtTop
    "E402",
    # LineTooLong
    "E501",
]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["troload"]

